+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.12 (unreleased)
+## 0.2.12
 
 - BREAKING CHANGE: Modified `catalystcenter_wireless_device_provision` resource to use the `/dna/intent/api/v1/wirelessControllers/%v/provision` API endpoint instead of `/dna/intent/api/v1/wireless/provision`
 - Add `assign_managed_ap_locations` resource